remove doubled function and move useVestNfts into global lib
