+++ Frontend-v1-Original/components/ssRewards/lib/queries.ts
@@ -22,8 +22,12 @@ import {
   VestNFT,
   Pair,
 } from "../../../stores/types/types";
-import { useGovToken, useVeToken, usePairs } from "../../../lib/global/queries";
-import { useVestNfts } from "../../ssVests/queries";
+import {
+  useGovToken,
+  useVeToken,
+  usePairs,
+  useVestNfts,
+} from "../../../lib/global/queries";
 
 const CANTO_OPTION_TOKEN = "0x9f9A1Aa08910867F38359F4287865c4A1162C202";
